extracting constants and code for provider-stream
